Added testing to the app

+++ app/controllers/staffs_controller.rb
@@ -32,8 +32,12 @@ class StaffsController < ApplicationController
   def destroy
     @staff = @taikai.staffs.find(params[:id])
 
-    if !@staff.destroy
-      flash[:alert] = "Unable to remove Staff #{@staff.display_name}"
+    if @staff.last_admin?
+      flash[:alert] = "Unable to remove Staff '#{@staff.display_name}', he is the last admin for taikai '#{@staff.taikai.shortname}'"
+    else
+      if !@staff.destroy
+        flash[:alert] = "Unable to remove Staff #{@staff.display_name}"
+      end
     end
     redirect_to controller: 'taikais', action: 'edit', id: @taikai
   end
